# Optimize Float.parse/1