@vanderhoop 2025-01-09
@dkuku 👏 Given your work with IO data in https://github.com/elixir-lang/elixir/pull/14130 and now in this PR... I'm wondering if you (or others) have ideas about other spots you think could benefit from similar optimizations?