@dkuku 2025-01-09
The language core is well optimized.
Even in cases where benchmarks show a 50% difference, we're only talking about nanoseconds in rare scenarios.